[file-renamed] Header.context.mock.ts Renamed file 'Header.context.mock.ts' to 'Header.context.test.mock.ts'. And also updated it's refrences in other files. I did so because we need a 'HeaderContext' mock for writing stories too.
