+++ components/layout/Navbar/Desktop/NavbarLinks.test.tsx
@@ -2,7 +2,7 @@ import { describe, expect, it, Mock, vi } from 'vitest';
 import { render } from '@testing-library/react';
 import NavbarLinks from '@/components/layout/Navbar/Desktop/NavbarLinks';
 import * as HeaderContextModule from '@/components/layout/Header/Header.context';
-import { mockUseHeaderContext } from '@/components/layout/Header/Header.context.mock';
+import { mockUseHeaderContext } from '@/components/layout/Header/Header.context.test.mock';
 
 describe('NavbarLinks', () => {
   it('should have multiple category links', () => {
